Set up files for testing
