+++ app.test.js
@@ -1,0 +1,16 @@
+const request = require('supertest');
+const app = require('./app');
+const environment = process.env.NODE_ENV || 'development';
+const configuration = require('./knexfile')[environment]
+const database = require('knex')(configuration)
+
+describe('Server', () => {
+  describe('init', () => {
+    it('should return a 200 status', async () => {
+      const res = await request(app).get('/')
+      expect(res.status).toBe(200)
+    })
+  })
+
+  describe()
+})
